tweak a few styles and move each story to its own file
