+++ .storybook/addons.js
@@ -1,2 +1,4 @@
 import '@storybook/addon-actions/register';
 import '@storybook/addon-links/register';
+import '@storybook/addon-info';
+import '@storybook/addon-backgrounds/register';
