+++ package.json
@@ -48,6 +48,8 @@
   },
   "devDependencies": {
     "@storybook/addon-actions": "^4.0.4",
+    "@storybook/addon-backgrounds": "^4.1.7",
+    "@storybook/addon-info": "^4.1.7",
     "@storybook/addon-links": "^4.0.4",
     "@storybook/addons": "^4.0.4",
     "@storybook/react": "^4.0.4",
