+++ src/stories/util.js
@@ -1,0 +1,10 @@
+import { theme } from "../lib/constants";
+
+export const backgroundOptions = [
+  { name: "light background", value: theme.lightGrayColor, default: true },
+  { name: "dark background", value: theme.darkestBlueColor }
+];
+
+
+export const testImageSrc =
+  "https://onelovemassive.com/wp-content/uploads/2016/10/test-image.png";
